refactor(fs): move uhyve check outside of init
This is effectively a style-related change that suppresses an info!

Apart from reducing unused code when running Hermit kernels in bare
metal or under QEMU, this change additionally moves the "is_uhyve"
checks to the caller function instead of the callee for reasons of
consistency.

Fixes https://github.com/hermit-os/kernel/issues/1969