style: improve user-facing kernel messages
- Move subjects to beginning of each message
- Unify certain messages across architecture-specific implementations
- State the component/task at the beginning of most messages
  (before verbs) to make log scrolling and pinpointing failures easier
- Grammar-related improvements
- Increase consistency of used wording
- In some cases, display relevant variables
- Replace debug! with error! before errors are returned.
- Replace some debug! instances with trace!
- Decrease the length of some sentences
- Make messages of standard actions (e.g. driver init) consistent
  (mostly.)
- Correct mentions of RTL8139 in the driver's messages
- Add more variables inside of placeholders

This change is not intended to be perfect and does not improve ALL
messages, with many changes being highly subjective.

It has a primary focus on debugging information and drivers.
The primary motivation for this change is mostly the high amount of
time spent trying to understand sequences of actions (and e.g. where
specifically the booting of a kernel may break), while also assisting
those that do not have a full picture of the kernel's source code.

Some messages were not modified to include the name of the component,
as it may be self-evident depending on the context and the debugging
activity (see: src/drivers/gem.rs) and increase the mental noise,
rather than decrease it.