fix(virtq/split): remove deadlocking Box around items
The split virtqueue implementation previously still used a Box around
the token ring items. The array the token ring is based on is howevery
already on the heap (in another Box). To avoid having to allocate for
individual items, only allocating the ring array up-front, remove the
Box around the items.

To successfully recover memory from the balloon device in an OOM
situation, virtqueue submission for a deflate request must not allocate.
During OOM handling the global allocator is already locked, so this
would deadlock.