feat(balloon): use talc fork for malloc without OOM handling
Talc previously did not support allocations without attempting to use
the OOM handler to recover. For the balloon device to allocate pages for
the balloon however we want to just probe for and then allocate a chunk
of pages. If the allocation fails we don't want to recover old memory,
we just want to allocate a smaller chunk (since recovery with the
balloon will be shrinking the balloon, causing us to uselessly shuffle
around pages, possibly even failing).

See https://github.com/SFBdragon/talc/pull/43
which is the PR for merging this modification upstream. Another API
function may be added for querying the current largest possible new
allocation in the future. In that case the balloon driver code should be
changed to use that instead of probing allocatable size in powers of
two.