refactor: use PageRangeAllocators instead of FREE_LISTs
This replaces uses of KERNEL_FREE_LIST with PageAlloc.
This replaces uses of PHYSICAL_FREE_LIST with FrameAlloc.

Note that this will slow down parts of x86-64 paging right now, since it is not possible to hold a global `FrameAlloc` lock anymore.
This means the lock has to be reacquired again and again until we improve the paging code or the `FrameAlloc` code.

Co-authored-by: m-mueller678 <m-mueller678@users.noreply.github.com>